backend/CRm/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-templated-mail==1.1.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flower==2.0.0
humanize==4.7.0
idna==3.4
importlib-metadata==4.12.0
kombu==5.3.1
oauthlib==3.2.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
zipp==3.16.2
